update seeing_is_believing
Former-commit-id: c2d29ddc16133667657650ef7f7294fafdd2e4e0

+++ app/lib/seeing_is_believing/evaluate_by_moving_files.rb
@@ -11,6 +11,7 @@
 # read the wrong file... of course, since we rewrite the file,
 # its body will be incorrect, anyway.
 
+require 'rbconfig'
 require 'timeout'
 require 'seeing_is_believing/error'
 require 'seeing_is_believing/result'
@@ -74,9 +75,11 @@ class SeeingIsBelieving
     end
 
     def set_back_to_initial_conditions
-      @was_backed_up ?
-        File.rename(backup_filename, filename) :
+      if @was_backed_up
+        File.rename(backup_filename, filename)
+      else
         File.delete(filename)
+      end
     end
 
     def write_program_to_file
@@ -97,32 +100,33 @@ class SeeingIsBelieving
       # setup environment variables
       env = ENV.to_hash.merge 'SIB_VARIABLES.MARSHAL.B64' =>
                                 [Marshal.dump(
-                                  event_stream_fd:   child_eventstream.to_i,
+                                  event_stream_fd:   4,
                                   max_line_captures: max_line_captures,
                                   num_lines:         program.lines.count,
                                   filename:          filename
                                 )].pack('m0')
 
       # evaluate the code in a child process
-      opts = { in:                  child_stdin,
-               out:                 child_stdout,
-               err:                 child_stderr,
-               child_eventstream => child_eventstream }
-      child = Process.detach Kernel.spawn(env, *popen_args, opts)
+      opts = {
+        4 =>    child_eventstream,
+        in:     child_stdin,
+        out:    child_stdout,
+        err:    child_stderr,
+        pgroup: true, # run it in its own process group so we can SIGINT the whole group
+      }
+      child_pid  = Kernel.spawn(env, *popen_args, opts)
+      child_pgid = Process.getpgid(child_pid)
 
       # close child streams b/c they won't emit EOF
       # until both child and parent references are closed
-      child_eventstream.close
-      child_stdout.close
-      child_stderr.close
-      child_stdin.close
+      close_streams(child_eventstream, child_stdout, child_stderr, child_stdin)
       stdin.sync = true
 
-      # send stdin
-      Thread.new {
+      # send stdin (char at a time b/c input could come from a stream)
+      Thread.new do
         provided_input.each_char { |char| stdin.write char }
         stdin.close
-      }
+      end
 
       # set up the event consumer
       consumer = EventStream::Consumer.new(events: eventstream, stdout: stdout, stderr: stderr)
@@ -130,26 +134,36 @@ class SeeingIsBelieving
 
       # wait for completion
       Timeout.timeout timeout_seconds do
-        exitstatus = child.value.exitstatus
-        consumer.process_exitstatus exitstatus
+        Process.wait child_pid
+        consumer.process_exitstatus($?.exitstatus)
         consumer_thread.join
       end
     rescue Timeout::Error
-      Process.kill "KILL", child.pid
       consumer.process_timeout timeout_seconds
-      consumer_thread.join # finish consuming events
     ensure
-      [stdin, stdout, stderr, eventstream].each { |io| io.close unless io.closed? }
+      allow_error(Errno::ESRCH)  { Process.kill "-INT", child_pgid } # negative makes it apply to the group
+      allow_error(Errno::ECHILD) { Process.wait child_pid } # I can't tell if this actually works, or just creates enough of a delay for the OS to finish cleaning up the thread
+      consumer_thread.join
+      close_streams(stdin, stdout, stderr, eventstream)
     end
 
     def popen_args
-      ['ruby-2.2.2',
+      [RbConfig.ruby,
          '-W0',                                     # no warnings (b/c I hijack STDOUT/STDERR)
          *(encoding ? ["-K#{encoding}"] : []),      # allow the encoding to be set
          '-I', File.expand_path('../..', __FILE__), # add lib to the load path
          *load_path_flags,                          # users can inject dirs to be added to the load path
          *require_flags,                            # users can inject files to be required
          filename]
+    end
+
+    def allow_error(error)
+      yield
+    rescue error
+    end
+
+    def close_streams(*streams)
+      streams.each { |io| io.close unless io.closed? }
     end
   end
 end
